Split different aggregations in their own files
