+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/aggregation/Aggregations.java
@@ -5,28 +5,8 @@
 
 package io.opentelemetry.sdk.metrics.aggregation;
 
-import io.opentelemetry.api.common.Labels;
-import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
-import io.opentelemetry.sdk.metrics.aggregator.AggregatorFactory;
-import io.opentelemetry.sdk.metrics.aggregator.CountAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.DoubleLastValueAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.DoubleMinMaxSumCountAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.DoubleSumAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.LongLastValueAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.LongMinMaxSumCountAggregator;
-import io.opentelemetry.sdk.metrics.aggregator.LongSumAggregator;
-import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
-import io.opentelemetry.sdk.metrics.common.InstrumentType;
-import io.opentelemetry.sdk.metrics.common.InstrumentValueType;
-import io.opentelemetry.sdk.metrics.data.MetricData;
-import io.opentelemetry.sdk.resources.Resource;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import javax.annotation.Nullable;
-import javax.annotation.concurrent.Immutable;
-
 public final class Aggregations {
+  private Aggregations() {}
 
   /**
    * Returns an {@code Aggregation} that calculates sum of recorded measurements.
@@ -34,7 +14,7 @@ public final class Aggregations {
    * @return an {@code Aggregation} that calculates sum of recorded measurements.
    */
   public static Aggregation sum() {
-    return Sum.INSTANCE;
+    return SumAggregation.INSTANCE;
   }
 
   /**
@@ -45,7 +25,7 @@ public final class Aggregations {
    *     recorded * measurements).
    */
   public static Aggregation count() {
-    return Count.INSTANCE;
+    return CountAggregation.INSTANCE;
   }
 
   /**
@@ -54,7 +34,7 @@ public final class Aggregations {
    * @return an {@code Aggregation} that calculates the last value of all recorded measurements.
    */
   public static Aggregation lastValue() {
-    return LastValue.INSTANCE;
+    return LastValueAggregation.INSTANCE;
   }
 
   /**
@@ -65,264 +45,6 @@ public final class Aggregations {
    * @return an {@code Aggregation} that calculates a simple summary of all recorded measurements.
    */
   public static Aggregation minMaxSumCount() {
-    return MinMaxSumCount.INSTANCE;
+    return MinMaxSumCountAggregation.INSTANCE;
   }
-
-  private enum MinMaxSumCount implements Aggregation {
-    INSTANCE;
-
-    @Override
-    public AggregatorFactory getAggregatorFactory(InstrumentValueType instrumentValueType) {
-      return instrumentValueType == InstrumentValueType.LONG
-          ? LongMinMaxSumCountAggregator.getFactory()
-          : DoubleMinMaxSumCountAggregator.getFactory();
-    }
-
-    @Override
-    public Accumulation merge(Accumulation a1, Accumulation a2) {
-      MinMaxSumCountAccumulation minMaxSumCountAccumulation1 = (MinMaxSumCountAccumulation) a1;
-      MinMaxSumCountAccumulation minMaxSumCountAccumulation2 = (MinMaxSumCountAccumulation) a2;
-      return MinMaxSumCountAccumulation.create(
-          minMaxSumCountAccumulation1.getCount() + minMaxSumCountAccumulation2.getCount(),
-          minMaxSumCountAccumulation1.getSum() + minMaxSumCountAccumulation2.getSum(),
-          Math.min(minMaxSumCountAccumulation1.getMin(), minMaxSumCountAccumulation2.getMin()),
-          Math.max(minMaxSumCountAccumulation1.getMax(), minMaxSumCountAccumulation2.getMax()));
-    }
-
-    @Override
-    public MetricData toMetricData(
-        Resource resource,
-        InstrumentationLibraryInfo instrumentationLibraryInfo,
-        InstrumentDescriptor descriptor,
-        Map<Labels, Accumulation> accumulationMap,
-        long startEpochNanos,
-        long epochNanos) {
-      if (accumulationMap.isEmpty()) {
-        return null;
-      }
-      List<MetricData.Point> points = getPointList(accumulationMap, startEpochNanos, epochNanos);
-      return MetricData.createDoubleSummary(
-          resource,
-          instrumentationLibraryInfo,
-          descriptor.getName(),
-          descriptor.getDescription(),
-          descriptor.getUnit(),
-          MetricData.DoubleSummaryData.create(points));
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName();
-    }
-  }
-
-  @Immutable
-  private enum Sum implements Aggregation {
-    INSTANCE;
-
-    @Override
-    public AggregatorFactory getAggregatorFactory(InstrumentValueType instrumentValueType) {
-      return instrumentValueType == InstrumentValueType.LONG
-          ? LongSumAggregator.getFactory()
-          : DoubleSumAggregator.getFactory();
-    }
-
-    @Override
-    public Accumulation merge(Accumulation a1, Accumulation a2) {
-      // TODO: Fix this by splitting the Aggregation per instrument value type.
-      if (a1 instanceof LongAccumulation) {
-        LongAccumulation longAccumulation1 = (LongAccumulation) a1;
-        LongAccumulation longAccumulation2 = (LongAccumulation) a2;
-        return LongAccumulation.create(longAccumulation1.getValue() + longAccumulation2.getValue());
-      }
-      DoubleAccumulation longAccumulation1 = (DoubleAccumulation) a1;
-      DoubleAccumulation longAccumulation2 = (DoubleAccumulation) a2;
-      return DoubleAccumulation.create(longAccumulation1.getValue() + longAccumulation2.getValue());
-    }
-
-    @Override
-    public MetricData toMetricData(
-        Resource resource,
-        InstrumentationLibraryInfo instrumentationLibraryInfo,
-        InstrumentDescriptor descriptor,
-        Map<Labels, Accumulation> accumulationMap,
-        long startEpochNanos,
-        long epochNanos) {
-      List<MetricData.Point> points = getPointList(accumulationMap, startEpochNanos, epochNanos);
-      boolean isMonotonic =
-          descriptor.getType() == InstrumentType.COUNTER
-              || descriptor.getType() == InstrumentType.SUM_OBSERVER;
-      return getSumMetricData(
-          resource, instrumentationLibraryInfo, descriptor, points, isMonotonic);
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName();
-    }
-  }
-
-  @Immutable
-  private enum Count implements Aggregation {
-    INSTANCE;
-
-    @Override
-    public AggregatorFactory getAggregatorFactory(InstrumentValueType instrumentValueType) {
-      return CountAggregator.getFactory();
-    }
-
-    @Override
-    public Accumulation merge(Accumulation a1, Accumulation a2) {
-      LongAccumulation longAccumulation1 = (LongAccumulation) a1;
-      LongAccumulation longAccumulation2 = (LongAccumulation) a2;
-      return LongAccumulation.create(longAccumulation1.getValue() + longAccumulation2.getValue());
-    }
-
-    @Override
-    public MetricData toMetricData(
-        Resource resource,
-        InstrumentationLibraryInfo instrumentationLibraryInfo,
-        InstrumentDescriptor descriptor,
-        Map<Labels, Accumulation> accumulationMap,
-        long startEpochNanos,
-        long epochNanos) {
-      List<MetricData.Point> points = getPointList(accumulationMap, startEpochNanos, epochNanos);
-
-      return MetricData.createLongSum(
-          resource,
-          instrumentationLibraryInfo,
-          descriptor.getName(),
-          descriptor.getDescription(),
-          "1",
-          MetricData.LongSumData.create(
-              /* isMonotonic= */ true, MetricData.AggregationTemporality.CUMULATIVE, points));
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName();
-    }
-  }
-
-  @Immutable
-  private enum LastValue implements Aggregation {
-    INSTANCE;
-
-    @Override
-    public AggregatorFactory getAggregatorFactory(InstrumentValueType instrumentValueType) {
-      return instrumentValueType == InstrumentValueType.LONG
-          ? LongLastValueAggregator.getFactory()
-          : DoubleLastValueAggregator.getFactory();
-    }
-
-    @Override
-    public Accumulation merge(Accumulation a1, Accumulation a2) {
-      // TODO: Define the order between accumulation.
-      return a2;
-    }
-
-    @Override
-    public MetricData toMetricData(
-        Resource resource,
-        InstrumentationLibraryInfo instrumentationLibraryInfo,
-        InstrumentDescriptor descriptor,
-        Map<Labels, Accumulation> accumulationMap,
-        long startEpochNanos,
-        long epochNanos) {
-      List<MetricData.Point> points = getPointList(accumulationMap, startEpochNanos, epochNanos);
-
-      switch (descriptor.getType()) {
-        case SUM_OBSERVER:
-          return getSumMetricData(
-              resource, instrumentationLibraryInfo, descriptor, points, /* isMonotonic= */ true);
-        case UP_DOWN_SUM_OBSERVER:
-          return getSumMetricData(
-              resource, instrumentationLibraryInfo, descriptor, points, /* isMonotonic= */ false);
-        case VALUE_OBSERVER:
-          return getGaugeMetricData(resource, instrumentationLibraryInfo, descriptor, points);
-        case COUNTER:
-        case UP_DOWN_COUNTER:
-        case VALUE_RECORDER:
-      }
-      return null;
-    }
-
-    @Override
-    public String toString() {
-      return getClass().getSimpleName();
-    }
-  }
-
-  @Nullable
-  private static MetricData getSumMetricData(
-      Resource resource,
-      InstrumentationLibraryInfo instrumentationLibraryInfo,
-      InstrumentDescriptor descriptor,
-      List<MetricData.Point> points,
-      boolean isMonotonic) {
-    switch (descriptor.getValueType()) {
-      case LONG:
-        return MetricData.createLongSum(
-            resource,
-            instrumentationLibraryInfo,
-            descriptor.getName(),
-            descriptor.getDescription(),
-            descriptor.getUnit(),
-            MetricData.LongSumData.create(
-                isMonotonic, MetricData.AggregationTemporality.CUMULATIVE, points));
-      case DOUBLE:
-        return MetricData.createDoubleSum(
-            resource,
-            instrumentationLibraryInfo,
-            descriptor.getName(),
-            descriptor.getDescription(),
-            descriptor.getUnit(),
-            MetricData.DoubleSumData.create(
-                isMonotonic, MetricData.AggregationTemporality.CUMULATIVE, points));
-    }
-    return null;
-  }
-
-  @Nullable
-  private static MetricData getGaugeMetricData(
-      Resource resource,
-      InstrumentationLibraryInfo instrumentationLibraryInfo,
-      InstrumentDescriptor descriptor,
-      List<MetricData.Point> points) {
-    switch (descriptor.getValueType()) {
-      case LONG:
-        return MetricData.createLongGauge(
-            resource,
-            instrumentationLibraryInfo,
-            descriptor.getName(),
-            descriptor.getDescription(),
-            descriptor.getUnit(),
-            MetricData.LongGaugeData.create(points));
-      case DOUBLE:
-        return MetricData.createDoubleGauge(
-            resource,
-            instrumentationLibraryInfo,
-            descriptor.getName(),
-            descriptor.getDescription(),
-            descriptor.getUnit(),
-            MetricData.DoubleGaugeData.create(points));
-    }
-    return null;
-  }
-
-  private static List<MetricData.Point> getPointList(
-      Map<Labels, Accumulation> accumulationMap, long startEpochNanos, long epochNanos) {
-    List<MetricData.Point> points = new ArrayList<>(accumulationMap.size());
-    accumulationMap.forEach(
-        (labels, aggregator) -> {
-          MetricData.Point point = aggregator.toPoint(startEpochNanos, epochNanos, labels);
-          if (point == null) {
-            return;
-          }
-          points.add(point);
-        });
-    return points;
-  }
-
-  private Aggregations() {}
 }
